requirements.txt
# Core dependencies
langchain>=0.1.0
openai>=1.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
fastapi>=0.100.0
uvicorn>=0.22.0
python-multipart>=0.0.6
datasets>=2.0.0

# ML dependencies
transformers>=4.30.0
torch>=2.0.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
chromadb>=0.4.22

# Development dependencies
pytest>=7.0.0
black>=23.0.0
isort>=5.12.0
mypy>=1.0.0
flake8>=6.0.0 